Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.strokeRect.clip-expected.txt

2d.strokeRect.clip
strokeRect is affected by clipping regions
Actual output:
Expected output:


PASS strokeRect is affected by clipping regions 
